Moved routes to routes/api sub dir
