+++ server/routes/api/conversations.js
@@ -1,6 +1,6 @@
 var express = require('express');
 var router = express.Router();
-var controller = require("../controllers/Conversations");
+var controller = require("../../controllers/Conversations");
 
 // Manage a single user's conversations
 router
