+++ server/routes/api/users.js
@@ -1,6 +1,6 @@
 var express = require('express');
 var router = express.Router();
-var controller = require("../controllers/Users");
+var controller = require("../../controllers/Users");
 
 // Manage a single user
 router
